Feat: Ticket, TicketInventory 책임 분리
